Changed marshaler funcs to take []byte
I made this change because it seems that it is very important
to set an accurate ContentLength value in the request. I
was running into some errors in the transport that came back
saying "Unsolicited response received on idle HTTP channel starting
with...". After switching to []byte for the marshalers, the
NewRequest method can accurately extract the ContentLength
by calling Len() on the bytes.Buffer. I'm going to change
it later to make it more generic again but for now this
works for me.